README.md
# MainDeveFolio

# Type
My personal webdevfolio using react.js and django. 

# Main Language
1. Javascript
2. Python.

# Frame-Works
- React (Client-Side): Used React for the UI dev
- Django (Server-Side): Used Django for back-end dev

# Tips and installation guild to use the framework/library
Install the followin as stated: 
1. Back-end Dev: django, DRF, corseheaders,
2. React, Scss. 

# How to use the folder
Requirement on how To use this you will have to navigate to the font-end base file and run "npm run dev". For the back-end,  activate the environment by running "venv/Scripts/activate" in the root folder then navigate to te back-end root folder and then run "python manage.py runserver".

# Main branch Link
https://github.com/Nelson19so/MainDeveFolio/

# Note
this Respository is for my personnal blog Development Website Portfolio and it is a prvate respositroy.